+++ tests/test_reports_events.py
@@ -32,7 +32,7 @@
 #  @author    Jeff Tanner <jefft@tune.com>
 #  @copyright 2014 Tune (http://www.tune.com)
 #  @license   http://opensource.org/licenses/MIT The MIT License (MIT)
-#  @version   $Date: 2014-11-03 15:19:08 $
+#  @version   $Date: 2014-11-19 07:02:45 $
 #  @link      https://developers.mobileapptracking.com @endlink
 #
 import unittest
@@ -113,8 +113,8 @@ class TestReportsEvents(unittest.TestCase):
             response = events.find(
                 self.__start_date,
                 self.__end_date,
-                filter="(status = 'approved')",
                 fields=events.fields(TUNE_FIELDS_RECOMMENDED),
+                filter="(status = 'approved')",
                 limit=10,
                 page=None,
                 sort={"created": "DESC"},
@@ -142,8 +142,8 @@ class TestReportsEvents(unittest.TestCase):
             response = events.export(
                 self.__start_date,
                 self.__end_date,
-                filter="(status = 'approved')",
                 fields=events.fields(TUNE_FIELDS_RECOMMENDED),
+                filter="(status = 'approved')",
                 format="csv",
                 response_timezone="America/Los_Angeles"
             )