+++ CHANGES.md
@@ -91,3 +91,9 @@ Version 0.9.14
 
 Beta public release on Novenber 06, 2014
 * Fixes to Cohort and Actuals reports
+
+Version 0.9.16
+--------------
+
+Beta public release on Novenber 19, 2014
+* Fixes to Cohort and Retention reports